Split automation test tiers and prevent late dumps
