+++ source/elektron/md/mdJucePlugin/mdAutomationFirmwareTest.cpp
@@ -1,113 +1,34 @@
-#include "mdController.h"
-#include "mdPluginProcessor.h"
+#include "mdAutomationTestSupport.h"
 
-#include "mdLib/mdautomation.h"
-#include "mdLib/mddevice.h"
-#include "juce_events/juce_events.h"
-
-#include <cstdlib>
 #include <iostream>
-#include <stdexcept>
 #include <string>
 
 namespace
 {
-	void require(const bool _condition, const std::string& _message)
-	{
-		if(_condition)
-			return;
-		throw std::runtime_error(_message);
-	}
+	using namespace mdAutomationTest;
 
-	void process(mdJucePlugin::AudioPluginAudioProcessor& _processor,
-		const int _blocks)
-	{
-		juce::AudioBuffer<float> audio(2, 128);
-		juce::MidiBuffer midi;
-		auto& audioProcessor = static_cast<juce::AudioProcessor&>(_processor);
-		for(int block = 0; block < _blocks; ++block)
-		{
-			audio.clear();
-			midi.clear();
-			audioProcessor.processBlock(audio, midi);
-			_processor.getController().processPendingMidiMessages();
-		}
-	}
-
-	struct MidiTelemetry
-	{
-		uint64_t consumed = 0;
-		size_t overflows = 0;
-	};
-
-	MidiTelemetry midiTelemetry(
-		mdJucePlugin::AudioPluginAudioProcessor& _processor)
-	{
-		MidiTelemetry result;
-		_processor.getPlugin().withDeviceLocked(
-			[&result](synthLib::Device* const _device)
-			{
-				if(const auto* const device = dynamic_cast<md::Device*>(_device))
-				{
-					result.consumed = device->getHardware().midiRxConsumedCount();
-					result.overflows = device->getHardware().midiRxOverflowCount();
-				}
-			});
-		return result;
-	}
-
-	bool waitForSynchronization(mdJucePlugin::AudioPluginAudioProcessor& _processor,
-		mdJucePlugin::Controller& _controller)
-	{
-		for(int attempt = 0; attempt < 2500; ++attempt)
-		{
-			process(_processor, 1);
-			if(_controller.isAutomationSynchronized())
-				return true;
-		}
-		return false;
-	}
-
-	bool hasLocalFirmware(mdJucePlugin::AudioPluginAudioProcessor& _processor)
-	{
-		return _processor.getPlugin().withDeviceLocked(
-			[](synthLib::Device* const _device)
-			{
-				return dynamic_cast<md::Device*>(_device) != nullptr;
-			});
-	}
-
-	pluginLib::Parameter& parameter(mdJucePlugin::Controller& _controller,
-		const md::MachineModel)
+	pluginLib::Parameter& parameter(mdJucePlugin::Controller& _controller)
 	{
 		auto* const result = _controller.getParameter("Level", 0);
 		require(result != nullptr, "test parameter was not registered");
 		return *result;
 	}
 
-	void hostWrite(pluginLib::Parameter& _parameter, const int _value)
-	{
-		_parameter.setValue(_parameter.getNormalisableRange().convertTo0to1(
-			static_cast<float>(_value)));
-	}
-
 	void verifyModel(const md::MachineModel _model)
 	{
-		mdJucePlugin::AudioPluginAudioProcessor processor(_model,
-			mdJucePlugin::AudioPluginAudioProcessor::EphemeralConfig{}, false);
-		auto& audioProcessor = static_cast<juce::AudioProcessor&>(processor);
-		audioProcessor.prepareToPlay(48000.0, 128);
-		if(!hasLocalFirmware(processor))
+		Harness harness(_model);
+		if(!harness.hasLocalFirmware())
 		{
 			std::cout << "mdAutomationFirmwareTest: SKIP "
-				<< (_model == md::MachineModel::Monomachine ? "MM" : "MD")
+				<< modelName(_model)
 				<< " (firmware unavailable)\n";
 			return;
 		}
+		harness.prepare();
 
-		auto& controller = dynamic_cast<mdJucePlugin::Controller&>(
-			processor.getController());
-		require(waitForSynchronization(processor, controller),
+		auto& audioProcessor = harness.audioProcessor;
+		auto& controller = harness.controller;
+		require(harness.synchronize(),
 			"initial firmware synchronization timed out (global "
 			+ std::to_string(controller.hasAutomationGlobalSnapshot()) + ", kit "
 			+ std::to_string(controller.hasAutomationKitSnapshot()) + ", base "
@@ -115,38 +36,38 @@ namespace
 		require(controller.getAutomationBaseChannel() < 16,
 			"firmware Global has MIDI base channel set to NONE");
 
-		auto& probe = parameter(controller, _model);
+		auto& probe = parameter(controller);
 		const auto beforeTransmitCount =
 			controller.getTransmittedAutomationChangeCount();
 
 		// Make the cache say zero without touching firmware, then require an explicit
 		// repeated/default host write to reach and be consumed by the firmware UART.
 		probe.setValueFromSynth(0, pluginLib::Parameter::Origin::PresetChange);
-		const auto beforeZero = midiTelemetry(processor);
+		const auto beforeZero = harness.telemetry();
 		hostWrite(probe, 0);
 		require(controller.getTransmittedAutomationChangeCount()
 			== beforeTransmitCount + 1,
 			"host write did not produce exactly one automation CC");
-		process(processor, 32);
-		const auto afterZero = midiTelemetry(processor);
+		harness.process(32);
+		const auto afterZero = harness.telemetry();
 		require(afterZero.consumed >= beforeZero.consumed + 3
 			&& afterZero.overflows == beforeZero.overflows,
 			"explicit zero automation CC was not consumed losslessly by firmware");
 
-		const auto beforeChanged = midiTelemetry(processor);
+		const auto beforeChanged = harness.telemetry();
 		hostWrite(probe, 127);
 		require(controller.getTransmittedAutomationChangeCount()
 			== beforeTransmitCount + 2,
 			"changed host write did not produce exactly one automation CC");
-		process(processor, 32);
-		const auto afterChanged = midiTelemetry(processor);
+		harness.process(32);
+		const auto afterChanged = harness.telemetry();
 		require(afterChanged.consumed >= beforeChanged.consumed + 3
 			&& afterChanged.overflows == beforeChanged.overflows,
 			"changed automation CC was not consumed losslessly by firmware");
 
 		const auto beforeStressCount =
 			controller.getTransmittedAutomationChangeCount();
-		const auto beforeStressMidi = midiTelemetry(processor);
+		const auto beforeStressMidi = harness.telemetry();
 		size_t expectedWrites = 0;
 		int ordinal = 0;
 		for(auto* const audioParameter : audioProcessor.getParameters())
@@ -165,32 +86,32 @@ namespace
 			hostWrite(*automationParameter, value);
 			++expectedWrites;
 			if((++ordinal & 31) == 0)
-				process(processor, 4);
+				harness.process(4);
 		}
-		process(processor, 64);
+		harness.process(64);
 		require(controller.getTransmittedAutomationChangeCount()
 			== beforeStressCount + expectedWrites,
 			"automation stress pass did not transmit every host write");
-		const auto afterStressMidi = midiTelemetry(processor);
+		const auto afterStressMidi = harness.telemetry();
 		require(afterStressMidi.consumed
 			>= beforeStressMidi.consumed + expectedWrites * 3
 			&& afterStressMidi.overflows == beforeStressMidi.overflows,
 			"automation stress stream was not consumed losslessly by firmware");
 
 		hostWrite(probe, 47);
-		process(processor, 32);
+		harness.process(32);
 		juce::MemoryBlock savedState;
 		audioProcessor.getStateInformation(savedState);
 		require(!savedState.isEmpty(), "processor produced empty persisted state");
 		hostWrite(probe, 12);
-		process(processor, 32);
+		harness.process(32);
 		require(probe.getUnnormalizedValue() == 12,
 			"pre-restore control write failed");
 		audioProcessor.setStateInformation(savedState.getData(),
 			static_cast<int>(savedState.getSize()));
 		const auto beforeRestoreFlush =
 			controller.getTransmittedAutomationChangeCount();
-		require(waitForSynchronization(processor, controller),
+		require(harness.synchronize(),
 			"state restore did not resynchronize automation");
 		require(probe.getUnnormalizedValue() == 47,
 			"firmware-backed automation value was not restored from DAW state");
@@ -198,9 +119,8 @@ namespace
 			>= beforeRestoreFlush + audioProcessor.getParameters().size(),
 			"DAW-state automation snapshot was not flushed back to firmware");
 
-		audioProcessor.releaseResources();
 		std::cout << "mdAutomationFirmwareTest: "
-			<< (_model == md::MachineModel::Monomachine ? "MM" : "MD")
+			<< modelName(_model)
 			<< " PASS\n";
 	}
 }
